Same treatment for cipher modes
