+++ src/lib/modes/aead/aead.h
@@ -55,6 +55,8 @@ class BOTAN_DLL AEAD_Mode : public Cipher_Mode
       * modes, and large enough that random collisions are unlikely).
       */
       size_t default_nonce_length() const override { return 12; }
+
+      virtual ~AEAD_Mode();
    };
 
 /**
