+++ src/lib/modes/aead/ccm/ccm.cpp
@@ -11,8 +11,6 @@
 
 namespace Botan {
 
-BOTAN_REGISTER_BLOCK_CIPHER_MODE_LEN2(CCM_Encryption, CCM_Decryption, 16, 3);
-
 /*
 * CCM_Mode Constructor
 */
