+++ src/lib/modes/aead/chacha20poly1305/chacha20poly1305.cpp
@@ -10,9 +10,6 @@
 
 namespace Botan {
 
-BOTAN_REGISTER_TRANSFORM_NOARGS(ChaCha20Poly1305_Encryption);
-BOTAN_REGISTER_TRANSFORM_NOARGS(ChaCha20Poly1305_Decryption);
-
 ChaCha20Poly1305_Mode::ChaCha20Poly1305_Mode() :
    m_chacha(make_stream_cipher("ChaCha")),
    m_poly1305(make_message_auth("Poly1305"))
