+++ src/lib/modes/aead/eax/eax.cpp
@@ -13,8 +13,6 @@
 
 namespace Botan {
 
-BOTAN_REGISTER_BLOCK_CIPHER_MODE_LEN(EAX_Encryption, EAX_Decryption, 0);
-
 namespace {
 
 /*
