+++ src/lib/modes/cfb/cfb.cpp
@@ -11,8 +11,6 @@
 
 namespace Botan {
 
-BOTAN_REGISTER_BLOCK_CIPHER_MODE_LEN(CFB_Encryption, CFB_Decryption, 0);
-
 CFB_Mode::CFB_Mode(BlockCipher* cipher, size_t feedback_bits) :
    m_cipher(cipher),
    m_feedback_bytes(feedback_bits ? feedback_bits / 8 : cipher->block_size())
