+++ src/lib/modes/ecb/ecb.cpp
@@ -10,19 +10,6 @@
 
 namespace Botan {
 
-template<typename T>
-Transform* make_ecb_mode(const Transform::Spec& spec)
-   {
-   std::unique_ptr<BlockCipher> bc(get_block_cipher(spec.arg(0)));
-   std::unique_ptr<BlockCipherModePaddingMethod> pad(get_bc_pad(spec.arg(1, "NoPadding")));
-   if(bc && pad)
-      return new T(bc.release(), pad.release());
-   return nullptr;
-   }
-
-BOTAN_REGISTER_TRANSFORM(ECB_Encryption, make_ecb_mode<ECB_Encryption>);
-BOTAN_REGISTER_TRANSFORM(ECB_Decryption, make_ecb_mode<ECB_Decryption>);
-
 ECB_Mode::ECB_Mode(BlockCipher* cipher, BlockCipherModePaddingMethod* padding) :
    m_cipher(cipher),
    m_padding(padding)
